feat(shopping-list-item): implement shoppingListItem actions

+++ src/Item/itemService.ts
@@ -1,27 +1,34 @@
-import { Item } from ".prisma/client";
 import { inject, injectable } from "inversify";
+import { IAssociationRepository } from "../common/associationRepository";
 import { ServiceResult, success } from "../common/ServiceResult";
 import { IValidator } from "../common/validator";
 import { TYPES } from "../modules/types";
+import { ShoppingListItemDto } from "../ShoppingListItem/shoppingListItemDto";
+import { ShoppingListItemModel } from "../ShoppingListItem/shoppingListItemModel";
 import { IRepository } from "./../common/repository";
 import { IService } from "./../common/service";
 import { ValidationErrorResult } from "./../common/validationResult";
 import { ItemDto } from "./itemDto";
+import { ItemModel } from "./itemModel";
 
 @injectable()
-export class ItemService implements IService<Item, ItemDto> {
-  itemRepository: IRepository<Item, ItemDto>;
+export class ItemService implements IService<ItemModel, ItemDto> {
+  itemRepository: IRepository<ItemModel, ItemDto>;
+  shoppingListItemRepository: IAssociationRepository<ShoppingListItemModel, ShoppingListItemDto>;
   itemValidator: IValidator<ItemDto>;
 
   constructor(
-    @inject(TYPES.IItemRepository) itemRepository: IRepository<Item, ItemDto>,
+    @inject(TYPES.IItemRepository) itemRepository: IRepository<ItemModel, ItemDto>,
+    @inject(TYPES.IShoppingListItemRepository)
+    shoppingListItemRepository: IAssociationRepository<ShoppingListItemModel, ShoppingListItemDto>,
     @inject(TYPES.IItemValidator) itemValidator: IValidator<ItemDto>
   ) {
     this.itemRepository = itemRepository;
+    this.shoppingListItemRepository = shoppingListItemRepository;
     this.itemValidator = itemValidator;
   }
 
-  async getAll(): Promise<ServiceResult<Item[]>> {
+  async getAll(): Promise<ServiceResult<ItemModel[]>> {
     const result = await this.itemRepository.getAll();
 
     if (!result.success) {
@@ -32,7 +39,7 @@ export class ItemService implements IService<Item, ItemDto> {
     return success(result.data);
   }
 
-  async get(id: number): Promise<ServiceResult<Item>> {
+  async get(id: number): Promise<ServiceResult<ItemModel>> {
     const result = await this.itemRepository.get(id);
 
     if (!result.success) {
@@ -42,7 +49,7 @@ export class ItemService implements IService<Item, ItemDto> {
     return success(result.data);
   }
 
-  async create(data: ItemDto): Promise<ServiceResult<Item>> {
+  async create(data: ItemDto): Promise<ServiceResult<ItemModel>> {
     const validationResult = this.itemValidator.validate(data);
 
     if (!validationResult.success) {
@@ -58,7 +65,7 @@ export class ItemService implements IService<Item, ItemDto> {
     return success(result.data);
   }
 
-  async update(id: number, data: ItemDto): Promise<ServiceResult<Item>> {
+  async update(id: number, data: ItemDto): Promise<ServiceResult<ItemModel>> {
     const validationResult = this.itemValidator.validate(data);
 
     if (!validationResult.success) {
@@ -74,7 +81,13 @@ export class ItemService implements IService<Item, ItemDto> {
     return success(result.data);
   }
 
-  async delete(id: number): Promise<ServiceResult<Item>> {
+  async delete(id: number): Promise<ServiceResult<boolean>> {
+    const associationResult = await this.shoppingListItemRepository.delete(null, id);
+
+    if (!associationResult.success) {
+      return associationResult;
+    }
+
     const result = await this.itemRepository.delete(id);
 
     if (!result.success) {
