+++ src/ShoppingList/ShoppingListService.ts
@@ -1,26 +1,33 @@
-import { ShoppingList } from ".prisma/client";
 import { inject, injectable } from "inversify";
 import { ServiceResult, success } from "../common/serviceResult";
 import { IValidator } from "../common/validator";
 import { TYPES } from "../modules/types";
+import { ShoppingListItemDto } from "../ShoppingListItem/shoppingListItemDto";
+import { ShoppingListItemModel } from "../ShoppingListItem/shoppingListItemModel";
+import { IAssociationRepository } from "./../common/associationRepository";
 import { IRepository } from "./../common/repository";
 import { IService } from "./../common/service";
 import { ShoppingListDto } from "./shoppingListDto";
+import { ShoppingListModel } from "./shoppingListModel";
 
 @injectable()
-export class ShoppingListService implements IService<ShoppingList, ShoppingListDto> {
-  shoppingListRepository: IRepository<ShoppingList, ShoppingListDto>;
+export class ShoppingListService implements IService<ShoppingListModel, ShoppingListDto> {
+  shoppingListRepository: IRepository<ShoppingListModel, ShoppingListDto>;
+  shoppingListItemRepository: IAssociationRepository<ShoppingListItemModel, ShoppingListItemDto>;
   shoppingListValidator: IValidator<ShoppingListDto>;
 
   constructor(
-    @inject(TYPES.IShoppingListRepository) shoppingListRepository: IRepository<ShoppingList, ShoppingListDto>,
+    @inject(TYPES.IShoppingListRepository) shoppingListRepository: IRepository<ShoppingListModel, ShoppingListDto>,
+    @inject(TYPES.IShoppingListItemRepository)
+    shoppingListItemRepository: IAssociationRepository<ShoppingListItemModel, ShoppingListItemDto>,
     @inject(TYPES.IShoppingListValidator) shoppingListValidator: IValidator<ShoppingListDto>
   ) {
     this.shoppingListRepository = shoppingListRepository;
+    this.shoppingListItemRepository = shoppingListItemRepository;
     this.shoppingListValidator = shoppingListValidator;
   }
 
-  async getAll(): Promise<ServiceResult<ShoppingList[]>> {
+  async getAll(): Promise<ServiceResult<ShoppingListModel[]>> {
     const result = await this.shoppingListRepository.getAll();
 
     if (!result.success) {
@@ -29,7 +36,8 @@ export class ShoppingListService implements IService<ShoppingList, ShoppingListD
 
     return success(result.data);
   }
-  async get(id: number): Promise<ServiceResult<ShoppingList>> {
+
+  async get(id: number): Promise<ServiceResult<ShoppingListModel>> {
     const result = await this.shoppingListRepository.get(id);
 
     if (!result.success) {
@@ -39,7 +47,7 @@ export class ShoppingListService implements IService<ShoppingList, ShoppingListD
     return success(result.data);
   }
 
-  async create(data: ShoppingListDto): Promise<ServiceResult<ShoppingList>> {
+  async create(data: ShoppingListDto): Promise<ServiceResult<ShoppingListModel>> {
     const validationResult = this.shoppingListValidator.validate(data);
 
     if (!validationResult.success) {
@@ -55,7 +63,7 @@ export class ShoppingListService implements IService<ShoppingList, ShoppingListD
     return success(result.data);
   }
 
-  async update(id: number, data: ShoppingListDto): Promise<ServiceResult<ShoppingList>> {
+  async update(id: number, data: ShoppingListDto): Promise<ServiceResult<ShoppingListModel>> {
     const validationResult = this.shoppingListValidator.validate(data);
 
     if (!validationResult.success) {
@@ -71,13 +79,19 @@ export class ShoppingListService implements IService<ShoppingList, ShoppingListD
     return success(result.data);
   }
 
-  async delete(id: number): Promise<ServiceResult<ShoppingList>> {
+  async delete(id: number): Promise<ServiceResult<boolean>> {
+    const associationResult = await this.shoppingListItemRepository.delete(id, null);
+
+    if (!associationResult.success) {
+      return associationResult;
+    }
+
     const result = await this.shoppingListRepository.delete(id);
 
     if (!result.success) {
       return result;
     }
 
-    return success(result.data);
+    return success(true);
   }
 }
